refactor(arch): move repositories to app/repositories
