+++ .agent/docs/qa/v1.20-qa.md
@@ -80,7 +80,7 @@
  
  **Question**: リポジトリの実装ファイルは `app/repositories/` と `app/utils/api/` のどちらに置くべきか？
  
- **Answer**: `app/utils/api/` に配置する。`app/repositories/` は使用しない。
+ **Answer**: `app/repositories/` に配置する。`app/repositories/` の方が直感的であるため、`app/utils/api/` は使用しない。
  
  **Decided on**: 2025-12-13
  **Related**: architecture.md
